REFACTOR test suite into 'Feature' & 'Unit' directories
